requirements.txt
Pillow
pytesseract
numpy
tk
re
